requirements.txt
AfricastalkingGateway==1.9
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.8.17
django-bootstrap-toolkit==2.15.0
django-bootstrap3==9.1.0
django-formtools==2.1
django-geojson==2.10.0
django-leaflet==0.22.0
django-phonenumber-field==1.3.0
django-rest-swagger==2.1.2
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
GDAL==2.2.2
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.3
olefile==0.44
openapi-codec==1.3.2
pandas==0.21.0
phonenumberslite==8.8.5
Pillow==4.3.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
simplejson==3.12.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
